Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@
 int main() {
     int a,b,c;
     scanf("%d %d %d", &a,&b,&c);
-    int avg=((a+b+c)/3);
+    float avg=((a+b+c)/3);
     printf("Average: %d", avg);
     return 0;
 }
